Adding an if-else statement to cover missing images
Introducing an if-else statement to cover the eventuality where an image is not present on disk. A check is made to see if the file exists on disk; if found, then a directory is created and the copied, else a warning message is logged, specifying the ImageID of the missing file and where it was expected to be found.

+++ PadChestTools/DataFilterer/Program.cs
@@ -51,8 +51,15 @@
                     string sourcePath = $"{inputDataDirectory}\\{image.imageDir}\\{image.imageID}";
                     string outputPath = $"{outputDataDirectory}\\{image.condition}\\{image.patientID}\\{image.projection}\\{image.imageID}";
 
-                    Directory.CreateDirectory($"{outputDataDirectory}\\{image.condition}\\{image.patientID}\\{image.projection}");
-                    File.Copy(sourcePath, outputPath);
+                    if (File.Exists(sourcePath))
+                    {
+                        Directory.CreateDirectory($"{outputDataDirectory}\\{image.condition}\\{image.patientID}\\{image.projection}");
+                        File.Copy(sourcePath, outputPath);
+                    }
+                    else
+                    {
+                        Console.WriteLine($"WARN: Could not find image {image.imageID} at {sourcePath}");
+                    }
                 }
                 Console.WriteLine($"Finished copying images for {conditionImagesSet.Key}");
             }
